Refactor autonomous routines to use constants for starting poses and feed time, and update moveAndShoot method signature for improved clarity

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/autos/AutoBuilder.java
@@ -86,7 +86,7 @@ public class AutoBuilder {
     }
 
     /* Shooter Actions */
-    public AutoBuilder moveAndShoot(double power, int feedTime, Pose2d targetPose) {
+    public AutoBuilder moveAndShoot(double power, double feedTime, Pose2d targetPose) {
         actions.add(new ParallelAction(
                         drive.actionBuilder(currentPose)
                                 .strafeToLinearHeading(
